# Optimize the layernorm operator with openmp for multi-threads.