--- a/paddle/fluid/operators/jit/more/intrinsic/layer_norm.cc
+++ b/paddle/fluid/operators/jit/more/intrinsic/layer_norm.cc
@@ -26,19 +26,13 @@ namespace intrinsic {
 void LayerNorm(float* x, float* out, float* mean, float* var,
                const float* scale, const float* bias, int height,
                const float epsilon, int right) {
-  __m256 sum;
-  __m256 mean_vec, var_vec;
-  __m128 hi, lo;
-  __m256 tmp;
-  size_t offset;
-  size_t j;
-  int block = YMM_FLOAT_BLOCK;
+  constexpr int block = YMM_FLOAT_BLOCK;
   const int rest = right % block;
-  const int end = right - rest;
+  const int num = right / block;
 
+  __m256 epsilon_vec = _mm256_set1_ps(epsilon);
   __m256 reverse_num_vec =
       _mm256_div_ps(_mm256_set1_ps(1.0), _mm256_set1_ps(right));
-  __m256 epsilon_vec = _mm256_set1_ps(epsilon);
   int rest_mask =
       ((-1) & (~((~0U) >> (sizeof(int) * 8 - (block - rest))))) & 0x0ff;
   __m256i mask_vec = _mm256_set_epi32(
@@ -47,109 +41,235 @@ void LayerNorm(float* x, float* out, float* mean, float* var,
       rest_mask & 0x8 ? 0xffffffff : 0, rest_mask & 0x4 ? 0xffffffff : 0,
       rest_mask & 0x2 ? 0xffffffff : 0, rest_mask & 0x1 ? 0xffffffff : 0);
 
-  for (int i = 0; i < height; ++i) {
-    offset = i * right;
+  void* buf = NULL;
+  size_t mean_buf_len = block * height;
+  size_t var_buf_len = mean_buf_len;
+  size_t rest_buf_len = mean_buf_len;
 
-    /* get mean */
-    sum = _mm256_setzero_ps();
-    for (j = offset; j < end + offset; j += block) {
-      sum = _mm256_add_ps(sum, _mm256_loadu_ps((const float*)x + j));
-    }
-    if (rest != 0) {
-      j = offset + right - block;
-      tmp = _mm256_loadu_ps((const float*)x + j);
-      tmp = _mm256_blendv_ps(_mm256_setzero_ps(), tmp,
-                             *(__m256*)&mask_vec);  // NOLINT
-      sum = _mm256_add_ps(sum, tmp);
-    }
-    hi = _mm256_extractf128_ps(sum, 1);
-    lo = _mm256_extractf128_ps(sum, 0);
-    sum = _mm256_add_ps(
-        sum, _mm256_insertf128_ps(
-                 _mm256_insertf128_ps(_mm256_setzero_ps(), hi, 0), lo, 1));
-    sum = _mm256_hadd_ps(sum, sum);
-    sum = _mm256_hadd_ps(sum, sum);
-    mean_vec = _mm256_mul_ps(sum, reverse_num_vec);
-    mean[i] = *reinterpret_cast<float*>(&mean_vec);
+  buf = malloc(sizeof(float) * (mean_buf_len + var_buf_len + rest_buf_len));
+  if (buf != NULL) {
+    float* rest_buf = NULL;
+    float* mean_buf = NULL;
+    float* var_buf = NULL;
+    mean_buf = reinterpret_cast<float*>(buf);
+    var_buf = mean_buf + mean_buf_len;
+    rest_buf = var_buf + var_buf_len;
 
-    /* get variance */
-    sum = _mm256_setzero_ps();
-    for (j = offset; j < end + offset; j += block) {
-      tmp = _mm256_sub_ps(_mm256_loadu_ps((const float*)x + j), mean_vec);
-      tmp = _mm256_mul_ps(tmp, tmp);
-      sum = _mm256_add_ps(sum, tmp);
-    }
-    if (rest != 0) {
-      j = offset + right - block;
-      tmp = _mm256_sub_ps(_mm256_loadu_ps((const float*)x + j), mean_vec);
-      tmp = _mm256_mul_ps(tmp, tmp);
-      tmp = _mm256_blendv_ps(_mm256_setzero_ps(), tmp,
-                             *(__m256*)&mask_vec);  // NOLINT
-      sum = _mm256_add_ps(sum, tmp);
-    }
-    hi = _mm256_extractf128_ps(sum, 1);
-    lo = _mm256_extractf128_ps(sum, 0);
-    sum = _mm256_add_ps(
-        sum, _mm256_insertf128_ps(
-                 _mm256_insertf128_ps(_mm256_setzero_ps(), hi, 0), lo, 1));
-    sum = _mm256_hadd_ps(sum, sum);
-    sum = _mm256_hadd_ps(sum, sum);
-    var_vec = _mm256_mul_ps(sum, reverse_num_vec);
-    var[i] = *reinterpret_cast<float*>(&var_vec);
+    memset(mean_buf, 0, mean_buf_len * sizeof(float));
+    memset(var_buf, 0, var_buf_len * sizeof(float));
 
-    /* get x_norm and calculate output*/
-    for (j = offset; j < end + offset; j += block) {
-      tmp = _mm256_sub_ps(_mm256_loadu_ps((const float*)x + j), mean_vec);
-      tmp = _mm256_div_ps(tmp,
-                          _mm256_sqrt_ps(_mm256_add_ps(var_vec, epsilon_vec)));
-      _mm256_storeu_ps(reinterpret_cast<float*>(out) + j, tmp);
+/* get mean */
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for collapse(2)
+#endif
+    for (int i = 0; i < height; ++i) {
+      for (int j = 0; j < num; j++) {
+        __m256 sum = _mm256_loadu_ps((const float*)mean_buf + i * block);
+        sum = _mm256_add_ps(
+            sum, _mm256_loadu_ps((const float*)x + i * right + j * block));
+        _mm256_storeu_ps(reinterpret_cast<float*>(mean_buf) + i * block, sum);
+      }
     }
+
     if (rest != 0) {
-      j = offset + right - block;
-      tmp = _mm256_sub_ps(_mm256_loadu_ps((const float*)x + j), mean_vec);
-      tmp = _mm256_div_ps(tmp,
-                          _mm256_sqrt_ps(_mm256_add_ps(var_vec, epsilon_vec)));
-      _mm256_storeu_ps(reinterpret_cast<float*>(out) + j, tmp);
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for
+#endif
+      for (int i = 0; i < height; ++i) {
+        int j = (i + 1) * right - block;
+        __m256 sum = _mm256_loadu_ps((const float*)mean_buf + i * block);
+        __m256 tmp = _mm256_loadu_ps((const float*)x + j);
+        tmp = _mm256_blendv_ps(_mm256_setzero_ps(), tmp,
+                               *(__m256*)&mask_vec);  // NOLINT
+        sum = _mm256_add_ps(sum, tmp);
+        _mm256_storeu_ps(reinterpret_cast<float*>(mean_buf) + i * block, sum);
+      }
+    }
+
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for
+#endif
+    for (int i = 0; i < height; ++i) {
+      __m256 sum = _mm256_loadu_ps((const float*)mean_buf + i * block);
+      __m128 hi = _mm256_extractf128_ps(sum, 1);
+      __m128 lo = _mm256_extractf128_ps(sum, 0);
+      sum = _mm256_add_ps(
+          sum, _mm256_insertf128_ps(
+                   _mm256_insertf128_ps(_mm256_setzero_ps(), hi, 0), lo, 1));
+      sum = _mm256_hadd_ps(sum, sum);
+      sum = _mm256_hadd_ps(sum, sum);
+      __m256 mean_vec = _mm256_mul_ps(sum, reverse_num_vec);
+      mean[i] = *reinterpret_cast<float*>(&mean_vec);
+      _mm256_storeu_ps(reinterpret_cast<float*>(mean_buf) + i * block,
+                       mean_vec);
+    }
+
+/* get variance */
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for collapse(2)
+#endif
+    for (int i = 0; i < height; ++i) {
+      for (int j = 0; j < num; j++) {
+        __m256 mean_vec = _mm256_loadu_ps((const float*)mean_buf + i * block);
+        __m256 sum = _mm256_loadu_ps((const float*)var_buf + i * block);
+        __m256 tmp = _mm256_sub_ps(
+            _mm256_loadu_ps((const float*)x + i * right + j * block), mean_vec);
+        tmp = _mm256_mul_ps(tmp, tmp);
+        sum = _mm256_add_ps(sum, tmp);
+
+        _mm256_storeu_ps(reinterpret_cast<float*>(var_buf) + i * block, sum);
+      }
+    }
+
+    if (rest != 0) {
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for
+#endif
+      for (int i = 0; i < height; ++i) {
+        int j = (i + 1) * right - block;
+        __m256 mean_vec = _mm256_loadu_ps((const float*)mean_buf + i * block);
+        __m256 sum = _mm256_loadu_ps((const float*)var_buf + i * block);
+        __m256 tmp =
+            _mm256_sub_ps(_mm256_loadu_ps((const float*)x + j), mean_vec);
+        tmp = _mm256_mul_ps(tmp, tmp);
+        tmp = _mm256_blendv_ps(_mm256_setzero_ps(), tmp,
+                               *(__m256*)&mask_vec);  // NOLINT
+        sum = _mm256_add_ps(sum, tmp);
+        _mm256_storeu_ps(reinterpret_cast<float*>(var_buf) + i * block, sum);
+      }
+    }
+
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for
+#endif
+    for (int i = 0; i < height; ++i) {
+      __m256 sum = _mm256_loadu_ps((const float*)var_buf + i * block);
+      __m128 hi = _mm256_extractf128_ps(sum, 1);
+      __m128 lo = _mm256_extractf128_ps(sum, 0);
+      sum = _mm256_add_ps(
+          sum, _mm256_insertf128_ps(
+                   _mm256_insertf128_ps(_mm256_setzero_ps(), hi, 0), lo, 1));
+      sum = _mm256_hadd_ps(sum, sum);
+      sum = _mm256_hadd_ps(sum, sum);
+      __m256 var_vec = _mm256_mul_ps(sum, reverse_num_vec);
+      var[i] = *reinterpret_cast<float*>(&var_vec);
+      _mm256_storeu_ps(reinterpret_cast<float*>(var_buf) + i * block, var_vec);
+    }
+
+/* get x_norm and calculate output*/
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for collapse(2)
+#endif
+    for (int i = 0; i < height; ++i) {
+      for (int j = 0; j < num; j++) {
+        __m256 mean_vec = _mm256_loadu_ps((const float*)mean_buf + i * block);
+        __m256 var_vec = _mm256_loadu_ps((const float*)var_buf + i * block);
+        __m256 tmp = _mm256_sub_ps(
+            _mm256_loadu_ps((const float*)x + i * right + j * block), mean_vec);
+        tmp = _mm256_div_ps(
+            tmp, _mm256_sqrt_ps(_mm256_add_ps(var_vec, epsilon_vec)));
+        _mm256_storeu_ps(reinterpret_cast<float*>(out) + i * right + j * block,
+                         tmp);
+      }
+    }
+
+    if (rest != 0) {
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for
+#endif
+      for (int i = 0; i < height; ++i) {
+        int j = (i + 1) * right - block;
+        __m256 mean_vec = _mm256_loadu_ps((const float*)mean_buf + i * block);
+        __m256 var_vec = _mm256_loadu_ps((const float*)var_buf + i * block);
+        __m256 tmp =
+            _mm256_sub_ps(_mm256_loadu_ps((const float*)x + j), mean_vec);
+        tmp = _mm256_div_ps(
+            tmp, _mm256_sqrt_ps(_mm256_add_ps(var_vec, epsilon_vec)));
+        _mm256_storeu_ps(reinterpret_cast<float*>(out) + j, tmp);
+      }
     }
 
     if (scale) {
       if (rest != 0) {
-        j = offset + right - block;
-        tmp = _mm256_loadu_ps((const float*)out + j);
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for
+#endif
+        for (int i = 0; i < height; ++i) {
+          int j = (i + 1) * right - block;
+          __m256 tmp = _mm256_loadu_ps((const float*)out + j);
+          _mm256_storeu_ps(reinterpret_cast<float*>(rest_buf) + i * block, tmp);
+        }
       }
-      for (j = offset; j < end + offset; j += block) {
-        _mm256_storeu_ps(
-            reinterpret_cast<float*>(out) + j,
-            _mm256_mul_ps(_mm256_loadu_ps((const float*)out + j),
-                          _mm256_loadu_ps((const float*)scale + j - offset)));
+
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for collapse(2)
+#endif
+      for (int i = 0; i < height; i++) {
+        for (int j = 0; j < num; j++) {
+          _mm256_storeu_ps(
+              reinterpret_cast<float*>(out) + i * right + j * block,
+              _mm256_mul_ps(
+                  _mm256_loadu_ps((const float*)out + i * right + j * block),
+                  _mm256_loadu_ps((const float*)scale + j * block)));
+        }
       }
+
       if (rest != 0) {
-        j = offset + right - block;
-        _mm256_storeu_ps(
-            reinterpret_cast<float*>(out) + j,
-            _mm256_mul_ps(tmp,
-                          _mm256_loadu_ps((const float*)scale + j - offset)));
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for
+#endif
+        for (int i = 0; i < height; ++i) {
+          int j = (i + 1) * right - block;
+          __m256 tmp = _mm256_loadu_ps((const float*)rest_buf + i * block);
+          _mm256_storeu_ps(
+              reinterpret_cast<float*>(out) + j,
+              _mm256_mul_ps(tmp,
+                            _mm256_loadu_ps((const float*)scale + j % right)));
+        }
       }
     }
 
     if (bias) {
       if (rest != 0) {
-        j = offset + right - block;
-        tmp = _mm256_loadu_ps((const float*)out + j);
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for
+#endif
+        for (int i = 0; i < height; ++i) {
+          int j = (i + 1) * right - block;
+          __m256 tmp = _mm256_loadu_ps((const float*)out + j);
+          _mm256_storeu_ps(reinterpret_cast<float*>(rest_buf) + i * block, tmp);
+        }
       }
-      for (j = offset; j < end + offset; j += block) {
-        _mm256_storeu_ps(
-            reinterpret_cast<float*>(out) + j,
-            _mm256_add_ps(_mm256_loadu_ps((const float*)out + j),
-                          _mm256_loadu_ps((const float*)bias + j - offset)));
+
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for collapse(2)
+#endif
+      for (int i = 0; i < height; i++) {
+        for (int j = 0; j < num; j++) {
+          _mm256_storeu_ps(
+              reinterpret_cast<float*>(out) + i * right + j * block,
+              _mm256_add_ps(
+                  _mm256_loadu_ps((const float*)out + i * right + j * block),
+                  _mm256_loadu_ps((const float*)bias + j * block)));
+        }
       }
+
       if (rest != 0) {
-        j = offset + right - block;
-        _mm256_storeu_ps(reinterpret_cast<float*>(out) + j,
-                         _mm256_add_ps(tmp, _mm256_loadu_ps((const float*)bias +
-                                                            j - offset)));
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for
+#endif
+        for (int i = 0; i < height; ++i) {
+          int j = (i + 1) * right - block;
+          __m256 tmp = _mm256_loadu_ps((const float*)rest_buf + i * block);
+          _mm256_storeu_ps(
+              reinterpret_cast<float*>(out) + j,
+              _mm256_add_ps(tmp,
+                            _mm256_loadu_ps((const float*)bias + j % right)));
+        }
       }
     }
+
+    free(buf);
   }
 }