Based on  [investigation of ERNIE](https://github.com/PaddlePaddle/benchmark/issues/180#issuecomment-546296907), we found that layernorm is lack of multi-threads JIT implemention. This PR is to add the layernorm multi-threads optimizing by using OpenMP.
But based on initial benchmark with ERNIE, single thread layernorm performance seems to be worse than before.

Total time | Baseline | Optimization
-- | -- | --
single thread | 1715.34 | 4624.7
multi-threads | 4009.16 | 2341.13